+++ src/ZendExpressive/Storageless/SessionPersistence.php
@@ -18,7 +18,7 @@
 
 declare(strict_types=1);
 
-namespace PSR7Sessions\Storageless\Session\Zend;
+namespace PSR7Sessions\ZendExpressive\Storageless;
 
 use Lcobucci\Clock\Clock;
 use Psr\Http\Message\ResponseInterface;
